Update 기타. 히든 넘버

+++ app/src/main/java/com/example/kotlincodingtest/baekjoon/기타/히든_넘버.kt
@@ -4,7 +4,7 @@ import java.io.BufferedReader
 
 fun main() = with(BufferedReader(System.`in`.bufferedReader())) {
     val n = readLine().toInt()
-    val hiddenNum = readLine().split(Regex("\\D+")).filter { it.isNotEmpty() }.map { it.toInt() }
+    val hiddenNum = readLine().split(Regex("\\D+")).filter { it.isNotEmpty() }.map { it.toLong() }
 
     println(hiddenNum.sum())
 }
